Add Cheese Interactables and Teleporter Class
This commit includes several changes to the codebase.
A new static method called 'CreateCheeseInteractable' has been added to the 'Interactable' class. This method allows for the creation of interactables that have a probability of having a random amount of cheese (1-3).

A new class called 'Teleporter' has been added, which is a subclass of 'Interactable'. This class allows for the creation of interactables that teleport the player to another room in the game.

Several new interactables have been added to rooms, and the probability of cheese for some of the interactables has been changed to create a more balanced and enjoyable gameplay experience.