Refactor File Structure for Improved Organization
This commit includes changes to the file structure of the project to improve organization and make it easier to navigate. The files have been reorganized into more logical groupings, which will make it easier for developers to find and modify specific code files. This refactoring will help to improve overall code maintainability and readability.