Add Abstract Enemy Class
This commit includes the addition of a new abstract class called 'Enemy'. This class will serve as a template for creating different types of enemies in the game. By creating an abstract class, a common interface is defined with common attributes and methods that all enemies will have, while still allowing for customization and specialization of individual enemy types. This new class will help to improve the overall design and scalability of the game.