Fix Typos and Interactable Errors in Red Kitchen Room
A new method called 'ScanRoom()' has been added to the 'Teleporter' class's 'Interact()' method. This method allows the player to scan the room after being teleported to a random location.

Typos have been fixed in the description of the 'group' Interactable, to ensure that the text is clear and accurate.

A missing exit/route has been added to the 'Red Kitchen' room, to ensure that players can navigate properly.

The red kitchen's Teleporter Interactable random locations have been fixed, to ensure that players are teleported to the intended locations.