feat: Add Max Currency and Fix Teleporter Bug
This commit includes the addition of a maximum currency limit in the game.

A bug was discovered where players could use the 'back' command to return to a previous room after using a teleporter. This bug has been fixed by creating a new stack upon using the teleporter. Additionally, a new checkpoint is created when the player uses a teleporter to ensure that progress is saved properly.